+++ packages/manager/src/features/Domains/DomainsLanding.tsx
@@ -2,19 +2,15 @@ import * as React from 'react';
 import { Domain } from '@linode/api-v4/lib/domains';
 import { useSnackbar } from 'notistack';
 import { useHistory, useLocation } from 'react-router-dom';
-import DomainIcon from 'src/assets/icons/entityIcons/domain.svg';
 import Button from 'src/components/Button';
 import { CircleProgress } from 'src/components/CircleProgress';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
-import Typography from 'src/components/core/Typography';
 import { DeletionDialog } from 'src/components/DeletionDialog/DeletionDialog';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import ErrorState from 'src/components/ErrorState';
 import LandingHeader from 'src/components/LandingHeader';
-import Link from 'src/components/Link';
 import Notice from 'src/components/Notice';
-import Placeholder from 'src/components/Placeholder';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import DisableDomainDialog from './DisableDomainDialog';
 import { Handlers as DomainHandlers } from './DomainActionMenu';
@@ -40,6 +36,7 @@ import { PaginationFooter } from 'src/components/PaginationFooter/PaginationFoot
 import Hidden from 'src/components/core/Hidden';
 import { CloneDomainDrawer } from './CloneDomainDrawer';
 import { EditDomainDrawer } from './EditDomainDrawer';
+import { DomainsEmptyLandingState } from './DomainsEmptyLandingPage';
 
 const DOMAIN_CREATE_ROUTE = '/domains/create';
 
@@ -207,35 +204,10 @@ export const DomainsLanding: React.FC<Props> = (props) => {
   if (domains?.results === 0) {
     return (
       <>
-        <DocumentTitleSegment segment="Domains" />
-        <Placeholder
-          title="Domains"
-          isEntity
-          icon={DomainIcon}
-          buttonProps={[
-            {
-              onClick: navigateToCreate,
-              children: 'Create Domain',
-            },
-            {
-              onClick: openImportZoneDrawer,
-              children: 'Import a Zone',
-            },
-          ]}
-        >
-          <Typography variant="subtitle1">
-            Create a Domain, add Domain records, import zones and domains.
-          </Typography>
-          <Typography variant="subtitle1">
-            <Link to="https://www.linode.com/docs/platform/manager/dns-manager-new-manager/">
-              Get help managing your Domains
-            </Link>
-            &nbsp;or&nbsp;
-            <Link to="https://www.linode.com/docs/">
-              visit our guides and tutorials.
-            </Link>
-          </Typography>
-        </Placeholder>
+        <DomainsEmptyLandingState
+          navigateToCreate={navigateToCreate}
+          openImportZoneDrawer={openImportZoneDrawer}
+        />
         <DomainZoneImportDrawer
           open={importDrawerOpen}
           onClose={closeImportZoneDrawer}